[FE] feat: DropDown 헤더 분리 (#29)
